a proto treelistview explorer with application in a file explorer
It compile and run with the tree and the list with basic functionality, but with some weird error, that shows the previous selected tree node and not the last in the list. Interesting it correctly expand the nodes. A problem in nana (tree widget report an incorrect selected item but correctly the expanded) or in the example?